Removed misc category from pages. Pages are added to this footer nav category automatically now if they have no valid categories.
Minor improvements to footer nav, code should be a bit cleaner. Also now we lowercase categories to prevent dupes and have some config variables that we can do filtering with.

'Misc' pages are now the last category on the footer nav and 'important' page are the first.

Added isDraft to pages that are drafts, we now filter these pages from the footer nav.
There is also a noFooter attribute that can be used to filter pages from the footer nav.

Updated video component to handle various youtube urls and covert them to proper embeds.
Eventually I'll add vimeo support too.
Renamed video component to Wiki-Video to avoid clashing with the html elm.

Added a sandangi video.